refactor: created parser file to better clean up lib.rs
